pyproject.toml
[tool.poetry]
name = "mydailyprop"
version = "0.1.0"
description = "A LangGraph + Reflex propaganda analyser"
license = "Apache 2.0"
authors = ["Pierre Guirriec <pierre.guirriec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
reflex = "^0.5.5"
langgraph = "^0.1.5"
langchain-openai = "^0.1.14"
firecrawl-py = "^0.0.20"
langchain-community = "^0.2.6"
langchainhub = "^0.1.20"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pillow = {version = "^10.4.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
